Move more RN components to screens
