+++ apps/native-component-list/src/screens/ExpoComponentsScreen.tsx
@@ -8,6 +8,9 @@ const screens = [
   'AdMob',
   'BarCodeScanner',
   'BlurView',
+  'ProgressViewIOS',
+  'ProgressBarAndroid',
+  'TouchableBounce',
   'Switch',
   'Slider',
   'Pressable',
